% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingApp.R
\name{runESS}
\alias{runESS}
\title{The Elementary Survey Sampling Shiny App}
\usage{
runESS(browser = T)
}
\arguments{
\item{browser}{logical: Option to open the app in the default system browser.}

\item{path}{string: Option to manually enter the directory path.}
}
\description{
This function runs the Elementary Survey Sampling Shiny app.
}
\examples{
# Runs the ESS Shiny app in the default system browser
runESS()

# Runs the ESS Shiny app in the RStudio IDE
runESS(browser = F)

}
\author{
Eric Brist
}
